Split configuration through annotation of the application
